fix(build): fix kernel compile and boot errors for QEMU run
Multiple issues prevented the kernel from compiling and booting:

- Cargo.toml: add instructions + abi_x86_interrupt features to x86_64
  crate; these gate x86_64::instructions module and HandlerFuncType impls
  needed by the IDT entry API
- src/lib.rs: add cfg_attr feature(abi_x86_interrupt) for bare-metal builds
- rust-toolchain.toml: add nightly toolchain pin with rust-src,
  llvm-tools-preview, x86_64-unknown-none so cargo build uses nightly
  (required for abi_x86_interrupt and build-std)
- src/gdt.rs: add kernel_data_segment to GDT and set DS/ES/FS/GS/SS
  to it; without a data segment the CPU faults when interrupt handlers
  try to access the stack segment; fix &raw const warning for mutable
  static reference
- src/interrupts.rs: fix IDT user interrupt entry indexing from usize
  to u8 (x86_64 0.15 IDT only implements Index<u8>)
- src/main.rs: remove redundant abi_x86_interrupt feature flag (covered
  by x86_64 crate feature now)

Kernel now boots cleanly in QEMU and reaches the cdk> serial console
with IDT active.

Made-with: Cursor

+++ src/interrupts.rs
@@ -70,9 +70,9 @@ pub fn init() {
                 .set_stack_index(crate::gdt::DOUBLE_FAULT_IST_INDEX);
         }
 
-        // Hardware interrupts
-        idt[usize::from(TIMER_INTERRUPT_ID)].set_handler_fn(timer_handler);
-        idt[usize::from(KEYBOARD_INTERRUPT_ID)].set_handler_fn(keyboard_handler);
+        // Hardware interrupts — IDT indexed by u8 in x86_64 0.15
+        idt[TIMER_INTERRUPT_ID].set_handler_fn(timer_handler);
+        idt[KEYBOARD_INTERRUPT_ID].set_handler_fn(keyboard_handler);
 
         idt
     });
